vue restructure
