+++ src/pages/integration-builder/builder/frameworks/vue/stepContent/buildingApp.mdx
